+++ src/bcra_sdk/models/__init__.py
@@ -1,4 +1,5 @@
 from .cheques import ResultGetChequesRechazadosV1
+from .denunciados import ResultGetChequeDenunciadoV1
 from .deudores import (
     EntidadHistorica,
     PeriodoHistorica,
@@ -10,6 +11,7 @@ from .entidades import ResultGetEntidadesV1
 __all__ = [
     "EntidadHistorica",
     "PeriodoHistorica",
+    "ResultGetChequeDenunciadoV1",
     "ResultGetChequesRechazadosV1",
     "ResultGetDeudasHistoricasV1",
     "ResultGetDeudasV1",